requirements.txt
anyio==3.6.1
click==8.1.3
fastapi==0.78.0
h11==0.13.0
idna==3.3
pydantic==1.9.1
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.2
anyio==3.6.1
azure-core==1.24.2
azure-identity==1.10.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
fastapi==0.78.0
google-api-core==2.8.2
google-auth==2.9.1
googleapis-common-protos==1.56.4
h11==0.13.0
idna==3.3
msal==1.18.0
msal-extensions==1.0.0
opencensus==0.10.0
opencensus-context==0.1.2
opencensus-ext-azure==1.1.5
portalocker==2.5.1
protobuf==4.21.2
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.10
uvicorn==0.18.2
anyio==3.6.1
azure-core==1.24.2
azure-identity==1.10.0
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cryptography==37.0.4
fastapi==0.78.0
google-api-core==2.8.2
google-auth==2.9.1
googleapis-common-protos==1.56.4
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
msal==1.18.0
msal-extensions==1.0.0
opencensus==0.10.0
opencensus-context==0.1.2
opencensus-ext-azure==1.1.5
portalocker==2.5.1
protobuf==4.21.2
psutil==5.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.1
PyJWT==2.4.0
requests==2.28.1
rsa==4.9
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.10
uvicorn==0.18.2
uvloop==0.16.0
